+++ activerecord/lib/active_record/attribute_methods/read.rb
@@ -11,10 +11,12 @@ module ActiveRecord
             ActiveModel::AttributeMethods::AttrNames.define_attribute_accessor_method(
               owner, name
             ) do |temp_method_name, attr_name_expr|
-              owner <<
-                "def #{temp_method_name}" <<
-                "  _read_attribute(#{attr_name_expr}) { |n| missing_attribute(n, caller) }" <<
-                "end"
+              owner.define_cached_method(name, as: temp_method_name, namespace: :active_record) do |batch|
+                batch <<
+                  "def #{temp_method_name}" <<
+                  "  _read_attribute(#{attr_name_expr}) { |n| missing_attribute(n, caller) }" <<
+                  "end"
+              end
             end
           end
       end